+++ crates/llms/Cargo.toml
@@ -17,7 +17,7 @@ async-stream.workspace = true
 async-trait.workspace = true
 aws-config.workspace = true
 aws-sdk-bedrockruntime = "1.55.0"
-aws-smithy-types = { version = "1.3.2" }             #, features = ["serde-serialize"] }
+aws-smithy-types = { version = "1.3.2" }
 bytes.workspace = true
 futures = { workspace = true }
 governor.workspace = true
@@ -36,6 +36,7 @@ tokio.workspace = true
 tracing.workspace = true
 tracing-futures.workspace = true
 util = { path = "../util" }
+
 ## For Chunking
 text-splitter = { git = "https://github.com/spiceai/text-splitter.git", rev = "58f9c21006e01e5e968c5de80a0398b3f5ec439a", features = [
     "markdown",
