+++ crates/runtime/src/model/embed.rs
@@ -22,7 +22,7 @@ use itertools::Itertools;
 use llms::HealthCheck;
 #[cfg(feature = "bedrock")]
 use llms::bedrock::{
-    self, BedrockClient,
+    self,
     embed::cohere::{CohereEmbeddingInputType, CohereEmbeddingTruncate, CohereEmbeddingType},
 };
